+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - [fix: fix lane router property name.](https://github.com/Tencent/spring-cloud-tencent/pull/1789)
 - [feat: support kafka lane.](https://github.com/Tencent/spring-cloud-tencent/pull/1765)
 - [fix: ApplicationContextAwareUtils may not be ready in postProcessAfterInitialization.](https://github.com/Tencent/spring-cloud-tencent/pull/1779)
+- [fix: kafka lane support dynamic lane tag.](https://github.com/Tencent/spring-cloud-tencent/pull/1784)